requirements.txt
accelerate
black==24.4.2
blobfile
codeshield
fairscale
fastapi
fire
flake8
httpx
hydra-core
hydra-zen
json-strong-typing
matplotlib
mesop
pandas
Pillow
pre-commit
pydantic==1.10.13
pydantic_core==2.18.2
python-dotenv
setuptools
tiktoken
torch
transformers
ufmt==2.7.0
usort==1.0.8
uvicorn
yfinance
zmq

llama_toolchain @ git+ssh://git@github.com/meta-llama/llama-toolchain.git
